# gh-117349: Micro-optimize a few `os.path` functions