Benchmarks:

- `ntpath`:
    ```shell
    # TODO: test isjunction() on Windows
    echo "isreserved()" && python -m timeit -s "import before.ntpath" "before.ntpath.isreserved('con')" && python -m timeit -s "import after.ntpath" "after.ntpath.isreserved('con')"
    echo "isreserved('.')" && python -m timeit -s "import before.ntpath" "before.ntpath.isreserved('.')" && python -m timeit -s "import after.ntpath" "after.ntpath.isreserved('.')"
    echo "expanduser()" && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~')" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~')"
    echo "realpath()"; python -m timeit -s "import test" "test.realpath1('.')"; python -m timeit -s "import test" "test.realpath2('.')"
    echo "realpath('nul')"; python -m timeit -s "import test" "test.realpath1('nul')"; python -m timeit -s "import test" "test.realpath2('nul')"
    ```

    ```none
    isreserved()
    200000 loops, best of 5: 1.66 usec per loop # before
    200000 loops, best of 5: 1.65 usec per loop # after
    # -> no difference
    isreserved('.')
    200000 loops, best of 5: 1.31 usec per loop # before
    500000 loops, best of 5: 966 nsec per loop # after
    # -> 1.36x faster (for . & ..)
    expanduser()
    200000 loops, best of 5: 1.72 usec per loop # before
    200000 loops, best of 5: 1.72 usec per loop # after
    # -> no difference
    realpath()
    2000 loops, best of 5: 132 usec per loop # before
    2000 loops, best of 5: 131 usec per loop # after
    # -> no difference
    realpath('nul')
    200000 loops, best of 5: 1.26 usec per loop # before
    500000 loops, best of 5: 907 nsec per loop # after
    # -> 1.39x faster (for nul)
    ```
- `posixpath`:
    ```shell
    # test.sh
    echo "ismount()" && python -m timeit -s "import before.posixpath" "before.posixpath.ismount('/Volumes/2GB_001')" && python -m timeit -s "import after.posixpath" "after.posixpath.ismount('/Volumes/2GB_001')"
    echo "expanduser()" && python -m timeit -s "import before.posixpath" "before.posixpath.expanduser('~')" && python -m timeit -s "import after.posixpath" "after.posixpath.expanduser('~')"
    echo "expanduser(b'~root')" && python -m timeit -s "import before.posixpath" "before.posixpath.expanduser(b'~root')" && python -m timeit -s "import after.posixpath" "after.posixpath.expanduser(b'~root')"
    echo "_normpath_fallback()" && python -m timeit -s "import before.posixpath" "before.posixpath._normpath_fallback('foo//bar')" && python -m timeit -s "import after.posixpath" "after.posixpath._normpath_fallback('foo//bar')"
    echo "abspath()" && python -m timeit -s "import before.posixpath" "before.posixpath.abspath('foo')" && python -m timeit -s "import after.posixpath" "after.posixpath.abspath('foo')"
    echo "abspath('/foo')" && python -m timeit -s "import before.posixpath" "before.posixpath.abspath('/foo')" && python -m timeit -s "import after.posixpath" "after.posixpath.abspath('/foo')"
    echo "realpath()" && python -m timeit -s "import before.posixpath" "before.posixpath.realpath('foo/../../..')" && python -m timeit -s "import after.posixpath" "after.posixpath.realpath('foo/../../..')"
    ```

    ```none
    ismount()
    10000 loops, best of 5: 20.3 usec per loop # before
    10000 loops, best of 5: 19.3 usec per loop # after
    # -> 1.05x faster
    expanduser()
    200000 loops, best of 5: 1.43 usec per loop # before
    200000 loops, best of 5: 1.42 usec per loop # after
    # -> no difference
    expanduser(b'~root')
    200000 loops, best of 5: 1.82 usec per loop # before
    200000 loops, best of 5: 1.75 usec per loop # after
    # -> 1.04x faster (for byte users)
    _normpath_fallback()
    200000 loops, best of 5: 1.07 usec per loop # before
    500000 loops, best of 5: 953 nsec per loop # after
    # -> 1.13x faster
    abspath()
    20000 loops, best of 5: 16.7 usec per loop # before
    20000 loops, best of 5: 16.5 usec per loop # after
    # -> no difference
    abspath('/foo')
    500000 loops, best of 5: 509 nsec per loop # before
    500000 loops, best of 5: 423 nsec per loop # after
    # -> 1.20x faster (for absolute paths)
    realpath()
    10000 loops, best of 5: 24.4 usec per loop # before
    10000 loops, best of 5: 23.9 usec per loop # after
    # -> 1.02x faster
    ```

<!-- gh-issue-number: gh-117349 -->
* Issue: gh-117349
<!-- /gh-issue-number -->