@nineteendo 2024-03-29
> Please revert all changes that do not actually have any impact on performance.

Does that include unnesting?