@nineteendo 2024-03-31
I think that covers everything. Thanks for all the feedback. I'll try to update the benchmarks as soon as possible.

_Note: I'm never trying to lint cpython without permission again, I'll leave that to my own repositories._