@AlexWaygood 2024-04-01
> * `ntpath.commonpath()` is roughly 256 nsec faster (which is 7%):

What version of Python are you measuring that with? A PGO-optimised build using the `main` branch? The speed of low-level things like list comprehensions is changing a lot with each Python version at the moment due to the work of the Faster CPython team, so it's important to do performance measurements with the `main` branch, especially with very localised micro-optimisations like those we're dealing with here