@nineteendo 2024-04-01
> What version of Python are you measuring that with?

It was 3.11, I incorrectly installed 3.12 and it wasn't on the path. I don't know how to built from the main branch (and I'm on my laptop now).