@AlexWaygood 2024-04-01
I'm also on my laptop -- I _only_ have a laptop ;)

List comprehensions are _much_ faster in Python 3.12+ than in Python <=3.11 due to the changes made as part of PEP-709, so that could easily account for the differences we're seeing between the benchmark results on my machine and yours, since the changes to `ntpath.commonpath()` involve changes to when and where list comprehensions are constructed.

You don't have to do so, but if you want to, there are instructions for building Python from source here: https://devguide.python.org/getting-started/setup-building/. (For accurate performance measurements, make sure you create a non-debug PGO-optimized build.)