@nineteendo 2024-04-01
> Also, it looks like you're making several changes to `ntpath.realpath()`, but I don't see a benchmark for that function in your PR description -- have you done one?

I couldn't test that on macOS, but I've justed tested it, and it's 1.38x faster for `nul` (in other cases the speedup will be smaller).

Also, I have updated the test for `ntpath.commonpath()` for 3.12, now it's only 2%.