@nineteendo 2024-04-01
I think that's everything, right? Or do you still want me to benchmark `ntpath.isjunction()`?