@AlexWaygood 2024-04-01
We've been discussing in the private Core Dev Discord server whether optimising code by replacing e.g. `if foo.startswith('////')` with `if foo[:4] == '////'`.

The underlying performance issue there seems to be that `str.startswith` is slower than it should be -- it would be _better_ if we could figure out why that is, and fix it so that it's just as fast as a slice-comparison, rather than going round changing all uses of `str.startswith` to use slice-comparisons instead. `str.startswith` seems significantly more idiomatic and readable -- and if we do improve the performance of `str.startswith` next week or next month, then we'll have muddied the git history in this module for no reason.