@AlexWaygood 2024-04-01
> Also, I have updated the test for `ntpath.commonpath()` for 3.12, now it's only 2%.

Using a PGO-optimized build from the `main` branch, here are my results using your benchmark for `ntpath.commonpath`:

```
(main) % ./python.exe -m timeit -s 'import ntpath' 'ntpath.commonpath(["foo/bar", "foo/baz"])'                                ~/dev/cpython
200000 loops, best of 5: 1.54 usec per loop
```

And with your PR branch:

```
(speedup-os.path) % ./python.exe -m timeit -s 'import ntpath' 'ntpath.commonpath(["foo/bar", "foo/baz"])'                     ~/dev/cpython
200000 loops, best of 5: 1.52 usec per loop
```

That's a 1.3% speedup, which is barely above the noise threshhold. I don't think it's worth changing that function for such a small performance improvement, unfortunately. Any change to CPython -- which is a _very_ mature, widely used piece of software -- carries a risk of breaking some other piece of code somewhere else.