@nineteendo 2024-04-01
> The underlying performance issue there seems to be that `str.startswith()` is slower than it should be -- it would be _better_ if we could figure out why that is, and fix it so that it's just as fast as a slice-comparison, rather than going round changing all uses of `str.startswith()` to use slice-comparisons instead.

I believe it should indeed be possible to speed up `str.startswith(prefix)`, but not `str.startswith((prefix1, prefix2))`, because there's not guarantee that the prefixes have the same length. I'll revert the optimisations for the first case.