@nineteendo 2024-04-01
The speed improvements seem way less impressive now. ;(
And I that's even without running my benchmark.