@nineteendo 2024-04-02
> With the latest version of your PR, I'm getting a reasonable slowdown on posixpath.ismount() from your PR branch.

You don't have a usb stick "2GB_001" plugged in, so the function will return `False` immediately after `s1 = os.lstat(path)`.
I believe something has been sped up on the main branch, so I just synced my branch.