@AlexWaygood 2024-04-02
> You don't have a usb stick "2GB_001" plugged in, so the function will return `False` immediately after `s1 = os.lstat(path)`.

heh, that makes sense