@nineteendo 2024-04-02
I've updated the benchmarks. Sadly, the only noticeable speedup for regular users is calling `posixpath.abspath()` with an absolute path.